+++ GerenciamentoFuncionarios/CalcSalario.java
@@ -1,0 +1,8 @@
+package ProjetoGestao;
+
+
+public interface CalcSalario{
+
+    public double CalcularSalario (double TaxaAumento, double salario);
+    
+}
