+++ GerenciamentoFuncionarios/ConsoleObserver.java
@@ -1,6 +1,9 @@
+package ProjetoGestao;
+
 public class ConsoleObserver implements Observer {
     @Override
     public void update(String message) {
         System.out.println("Notificação: " + message);
     }
 }
+
