+++ GerenciamentoFuncionarios/Main.java
@@ -1,3 +1,5 @@
+package ProjetoGestao;
+
 import java.util.Scanner;
 
 public class Main {
@@ -21,22 +23,27 @@ public class Main {
 
             switch (opcao) {
                 case 1:
+                System.out.println("");
                 System.out.print("Nome: ");
-                String nome = scanner.nextLine();
+                String nome = scanner.next();
                 System.out.print("Idade: ");
                 int idade = scanner.nextInt();
                 System.out.print("Salário: ");
                 double salario = scanner.nextDouble();
                 System.out.print("Cargo: ");
-                String cargo = scanner.nextLine();
-                scanner.nextLine(); 
+                String cargo = scanner.next();
+                
                 funcionario = new Funcionarios(nome, idade, salario, cargo);
                 empresa.addFuncionario(funcionario);
+                System.out.println();
                 break;
                 case 2:
+                	System.out.println("");
                     System.out.println(empresa.toString());
+                    System.out.println("");
                 break;
                 case 3:
+                	System.out.println("");
                     System.out.print("Qual o número correspondente do funcionário: ");
                     int funcionarioIndice = scanner.nextInt();
                     Funcionarios funcionarioAdd = empresa.getFuncionario(funcionarioIndice);
@@ -44,12 +51,15 @@ public class Main {
                     System.out.print("Qual a taxa de aumento desejada: ");
                     double TaxaAumento = scanner.nextDouble();
                     funcionarioAdd.CalcSalario(TaxaAumento);
+                    System.out.println("");
                     break;
                 case 4:
                     System.out.println("Saindo...");
                     break;
                 default:
+                	System.out.println("");
                     System.out.println("Opção inválida!");
+                    System.out.println("");
             }
