+++ GerenciamentoFuncionarios/Observer.java
@@ -1,3 +1,6 @@
+package ProjetoGestao;
+
 public interface Observer {
     void update(String message);
 }
+
