+++ GerenciamentoFuncionarios/Subject.java
@@ -1,6 +1,9 @@
+package ProjetoGestao;
+
 
 public interface Subject {
     void addObserver(Observer observer);  
     void removeObserver(Observer observer); 
     void notifyObservers(String message); 
 }
+
