--- GerenciamentoFuncionarios/src/ProjetoGestao/AumentoSalario.java
@@ -1,10 +1,0 @@
-public class AumentoSalario implements CalcSalario {
-
-    @Override
-    public double CalcSalario(double TaxaAumento, double salario) {
-        System.out.println("Salário Atualizado!");
-        return  salario = salario * TaxaAumento;
-    }
-}
-    
-
