--- GerenciamentoFuncionarios/src/ProjetoGestao/CalcSalario.java
@@ -1,5 +1,0 @@
-public interface CalcSalario{
-
-    public double CalcSalario (double TaxaAumento, double salario);
-    
-}
